src/requirementsPMO.txt
diff-match-patch==20121119
Django==1.8
django-crispy-forms==1.6.1
django-import-export==1.0.0
django-registration-redux==1.4
et-xmlfile==1.0.1
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.1
Pillow==4.3.0
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
